New top bar UI - WIP
