+++ packages/webui/src/client/ui/ClockView/CameraScreen/Part.tsx
@@ -7,7 +7,7 @@ import { PieceExtended } from '../../../lib/RundownResolver.js'
 import { getAllowSpeaking, getAllowVibrating } from '../../../lib/localStorage.js'
 import { getPartInstanceTimingValue } from '../../../lib/rundownTiming.js'
 import { AutoNextStatus } from '../../RundownView/RundownTiming/AutoNextStatus.js'
-import { CurrentPartOrSegmentRemaining } from '../../RundownView/RundownTiming/CurrentPartOrSegmentRemaining.js'
+import { CurrentPartOrSegmentRemaining } from '../../RundownView/RundownHeader/CurrentPartOrSegmentRemaining.js'
 import { PartCountdown } from '../../RundownView/RundownTiming/PartCountdown.js'
 import { PartDisplayDuration } from '../../RundownView/RundownTiming/PartDuration.js'
 import { TimingDataResolution, TimingTickResolution, useTiming } from '../../RundownView/RundownTiming/withTiming.js'
