+++ packages/webui/src/client/ui/RundownView/RundownHeader/Countdown.scss
@@ -1,0 +1,24 @@
+@import '../../../styles/colorScheme';
+
+.countdown {
+	display: flex;
+	align-items: baseline;
+	justify-content: space-between;
+	gap: 0.6em;
+	color: rgba(255, 255, 255, 0.6);
+	transition: color 0.2s;
+
+	&__label {
+		font-size: 0.7em;
+		font-weight: 600;
+		letter-spacing: 0.1em;
+		text-transform: uppercase;
+		white-space: nowrap;
+	}
+
+	&__value {
+		font-size: 1.4em;
+		font-variant-numeric: tabular-nums;
+		letter-spacing: 0.05em;
+	}
+}
