+++ packages/webui/src/client/ui/SegmentList/LinePart.tsx
@@ -7,7 +7,7 @@ import { contextMenuHoldToDisplayTime } from '../../lib/lib.js'
 import { RundownUtils } from '../../lib/rundown.js'
 import { getElementDocumentOffset } from '../../utils/positions.js'
 import { IContextMenuContext } from '../RundownView.js'
-import { CurrentPartOrSegmentRemaining } from '../RundownView/RundownTiming/CurrentPartOrSegmentRemaining.js'
+import { CurrentPartOrSegmentRemaining } from '../RundownView/RundownHeader/CurrentPartOrSegmentRemaining.js'
 import { PieceUi, SegmentUi } from '../SegmentContainer/withResolvedSegment.js'
 import { SegmentTimelinePartElementId } from '../SegmentTimeline/Parts/SegmentTimelinePart.js'
 import { LinePartIdentifier } from './LinePartIdentifier.js'
