+++ packages/webui/src/client/ui/Shelf/PartTimingPanel.tsx
@@ -8,7 +8,7 @@ import { DBRundownPlaylist } from '@sofie-automation/corelib/dist/dataModel/Rund
 import { dashboardElementStyle } from './DashboardPanel.js'
 import { RundownLayoutsAPI } from '../../lib/rundownLayouts.js'
 import { getAllowSpeaking, getAllowVibrating } from '../../lib/localStorage.js'
-import { CurrentPartOrSegmentRemaining } from '../RundownView/RundownTiming/CurrentPartOrSegmentRemaining.js'
+import { CurrentPartOrSegmentRemaining } from '../RundownView/RundownHeader/CurrentPartOrSegmentRemaining.js'
 import { CurrentPartElapsed } from '../RundownView/RundownTiming/CurrentPartElapsed.js'
 import { getIsFilterActive } from '../../lib/rundownLayouts.js'
 import { UIShowStyleBase } from '@sofie-automation/meteor-lib/dist/api/showStyles'
